codedocgen/requirements.txt
pyautogen>=0.2.0
requests>=2.25.0
questionary>=1.10.0
pyyaml>=6.0
colorama>=0.4.4
markdown>=3.3.0
ollama>=0.1.0
fix-busted-json>=0.0.18